requirements.txt
Django==2.1.7
django-crispy-forms==1.7.2
django-filter==2.1.0
djangorestframework==3.9.2
ez-setup==0.9
Markdown==3.1
mysql==0.0.2
mysqlclient==1.4.2.post1
Pillow==5.4.1
pytz==2018.9